+++ packages/rsc/examples/basic/src/server.tsx
@@ -1,4 +1,4 @@
-import { renderRequest } from "@hiogawa/vite-rsc/extra/server";
+import { renderRequest } from "@hiogawa/vite-rsc/extra/rsc";
 import {
   changeServerCounter,
   resetServerCounter,
