+++ packages/rsc/tsdown.config.ts
@@ -9,9 +9,9 @@ export default defineConfig({
     "src/core/browser.ts",
     "src/core/ssr.ts",
     "src/core/rsc.ts",
-    "src/extra/server.tsx",
-    "src/extra/ssr.tsx",
     "src/extra/browser.tsx",
+    "src/extra/ssr.tsx",
+    "src/extra/rsc.tsx",
   ],
   format: ["esm"],
   external: [/^virtual:/, /^@hiogawa\/vite-rsc\//],
